refactor: loginService 기능 userService로 통합
